requirements.txt
requests
python-dotenv
loguru
matplotlib
numpy
